Update documentation, and organization
